refactor: consolidate parseRouteParams and route functions into a single file;
fix: issue where axios instance wasn't being passed to the route function;

feat: pass custom axios instances to routeFactory class;

docs:  update readme;
