+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@eddielee394/axios-route-params-utils",
-  "version": "0.1.0",
+  "version": "0.1.5",
   "description": "Axios utility functions to automatically handle extracting dynamic route parameters from url segments",
   "main": "build/index.js",
   "scripts": {
