+++ src/lib/routeFactory.js
@@ -1,0 +1,11 @@
+import { getRoute } from "./route";
+
+class RouteFactory {
+  constructor(axiosInstance) {
+    this.instance = axiosInstance;
+  }
+
+  route = (path = "") => getRoute(path, this.instance);
+}
+
+export default RouteFactory;
