--- src/lib/tests/parseRouteParams.test.js
@@ -1,7 +1,0 @@
-import parseRouteParams from "../parseRouteParams";
-
-describe("parseRoutePrams tests", () => {
-  it("should be a function", () => {
-    expect(parseRouteParams).toEqual(expect.any(Function));
-  });
-});
